Standardizing types both inside dbscan and in the sparse csr prims

+++ cpp/src/dbscan/dbscan_api.cpp
@@ -19,7 +19,7 @@
 #include "dbscan.hpp"
 
 cumlError_t cumlSpDbscanFit(cumlHandle_t handle, float *input, long n_rows,
-                            long n_cols, float eps, int min_pts, int *labels,
+                            long n_cols, float eps, int min_pts, long *labels,
                             size_t max_bytes_per_batch, int verbose) {
   cumlError_t status;
   ML::cumlHandle *handle_ptr;
@@ -43,9 +43,8 @@ cumlError_t cumlSpDbscanFit(cumlHandle_t handle, float *input, long n_rows,
 }
 
 cumlError_t cumlDpDbscanFit(cumlHandle_t handle, double *input, long n_rows,
-                            long n_cols, double eps, int min_pts,
-                            int *labels, size_t max_bytes_per_batch,
-                            int verbose) {
+                            long n_cols, double eps, int min_pts, long *labels,
+                            size_t max_bytes_per_batch, int verbose) {
   cumlError_t status;
   ML::cumlHandle *handle_ptr;
   std::tie(handle_ptr, status) = ML::handleMap.lookupHandlePointer(handle);
